+++ src/editor/src/editor/plugin/fontFamily/index.js
@@ -5,7 +5,7 @@
 KISSY.add("editor/plugin/fontFamily/index", function (S, Editor, ui, cmd) {
 
     function FontFamilyPlugin(config) {
-this.config=config||{};
+        this.config = config || {};
     }
 
     S.augment(FontFamilyPlugin, {
@@ -15,9 +15,10 @@ this.config=config||{};
 
             var fontFamilies = this.config;
 
-            fontFamilies = fontFamilies || {};
+            var menu = {};
 
-            S.mix(fontFamilies, {
+
+            S.mix(menu, {
                 children:[
                     //ie 不认识中文？？？
                     {
@@ -70,9 +71,9 @@ this.config=config||{};
                     }
                 ],
                 width:"130px"
-            }, false);
+            });
 
-            S.each(fontFamilies.children, function (item) {
+            S.each(menu.children, function (item) {
                 var attrs = item.elAttrs || {},
                     value = item.value;
                 attrs.style = attrs.style || "";
@@ -80,16 +81,14 @@ this.config=config||{};
                 item.elAttrs = attrs;
             });
 
-            editor.addSelect("fontFamily", {
+            fontFamilies.menu = S.mix(menu, fontFamilies.menu);
+
+            editor.addSelect("fontFamily", S.mix({
                 cmdType:"fontFamily",
                 defaultCaption:"字体",
                 width:130,
-                mode:Editor.WYSIWYG_MODE,
-                menu:{
-                    width:fontFamilies.width,
-                    children:fontFamilies.children
-                }
-            }, ui.Select);
+                mode:Editor.WYSIWYG_MODE
+            }, fontFamilies), ui.Select);
         }
     });